feat: Move Atlas Blueprint content to own file
